MAILBOX-359 POJOify MessageMetaData & remove SimpleMessageMetaData
