+++ mailbox/store/src/test/java/org/apache/james/mailbox/store/AbstractCombinationManagerTest.java
@@ -384,7 +384,9 @@ public abstract class AbstractCombinationManagerTest {
         assertThat(messageManager1.getMessages(MessageRange.all(), FetchGroupImpl.MINIMAL, session)).isEmpty();
         assertThat(messageManager2.getMessages(MessageRange.all(), FetchGroupImpl.MINIMAL, session))
             .hasSize(1)
-            .extractingResultOf("getMessageId").containsOnly(messageId);
+            .extractingResultOf("messageMetaData")
+            .extractingResultOf("getMessageId")
+            .containsOnly(messageId);
     }
 
     @Test
